backend/requirements.txt
fastapi>=0.121.3
uvicorn>=0.38.0
pandas>=2.3.3
numpy>=2.3.5
scikit-learn>=1.7.2
python-multipart>=0.0.20
sqlalchemy>=2.0.44
groq>=0.36.0
python-dotenv>=1.2.1
reportlab>=4.4.5
matplotlib>=3.10.7
seaborn>=0.13.2
requests>=2.32.5
pyarrow>=22.0.0
openpyxl>=3.1.5
aiofiles>=25.1.0
pytest>=9.0.1
httpx>=0.28.1
plotly>=5.24.1
scipy>=1.15.2
joblib>=1.4.2
xgboost>=2.1.3
lightgbm>=4.5.0
optuna>=4.1.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.12.5
resend>=2.19.0
alembic>=1.17.2
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
sentry-sdk[fastapi]>=2.0.0